requirements.txt
streamlit
pandas
numpy
scikit-learn
torch
shap
plotly
python-dotenv
bcrypt
authlib
httpx